Reorganize resolvers
