+++ src/resolver/Mutation/remove.js
