+++ src/resolver/Query/single.js
